--- demos/src/element-tracking.mustache
@@ -1,15 +1,0 @@
-<div class="info">
-	<dl>
-		<dt>Total Elements</dt>
-		<dd id="total"></dd>
-		<dt>In View Elements</dt>
-		<dd id="inView"></dd>
-		<dt>On Screen Elements</dt>
-		<dd id="onScreen"></dd>
-	</dl>
-	<label for="listen">Track Elements</label>
-	<input type="checkbox" name="listen" value="on" checked />
-	<label for="threshold">In View Threshold <span id="threshold">100</span>%</label>
-	<input type="range" name="threshold" min="10" value="100" max="100" step="10" />
-</div>
-<div data-o-viewport-track=""></div>
